Resolve handling of default branch; refactor.
In the context of git-branches, there is no such thing as a default
branch (as in, the HEAD of remote), because in general case, it's
unclear if origin remote is the canonical remote, and it may or may not
exist. So just use the provided base branch only.

Refactor to simplify code. Now that it's internal, no need for options
struct.

Also rename `vcs` local variables to `git`, since git-branches only
ever deals with git repos. This makes it clearer that we can be
git-specific when it comes to handling of default branches, etc.