Hide trashed ("trash/" prefix) branches by default.
This makes it possible to move branches into a "trash". That way, the
branch is out of the way and doesn't interfere with your active
branches. At the same time, it's still easy to recover if you happen to
need it by using -all flag to list all branches.

This makes it possible to delete unneeded branches more eagerly,
without waiting until you're more certain you won't need a branch again.

It's similar to the ability to move files to a trash. Having that
ability makes it easier and less costly to delete files early, because
they're more recoverable. Apply same principle to git branches.